research/lucineer_analysis/lucineer/download/ternaryair/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ternaryair"
version = "1.0.0"
description = "Open Source Air-Gapped Ternary Inference Hardware SDK"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Casey DiGennaro", email = "casey@superinstance.ai"}
]
maintainers = [
    {name = "Casey DiGennaro", email = "casey@superinstance.ai"}
]
keywords = [
    "ai",
    "machine-learning",
    "inference",
    "ternary",
    "edge-ai",
    "hardware",
    "privacy",
    "security",
    "llm",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Topic :: Hardware",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "pyusb>=1.2.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
]
training = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]
all = [
    "ternaryair[dev,server,training,docs]",
]

[project.scripts]
ternaryair = "ternaryair.cli:main"

[project.urls]
Homepage = "https://github.com/superinstance/ternaryair"
Documentation = "https://github.com/superinstance/ternaryair#readme"
Repository = "https://github.com/superinstance/ternaryair.git"
Issues = "https://github.com/superinstance/ternaryair/issues"
Changelog = "https://github.com/superinstance/ternaryair/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["software/sdk"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docs
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"

[[tool.mypy.overrides]]
module = "usb.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow",
    "hardware: requires actual hardware",
    "simulator: uses software simulator",
]

[tool.coverage.run]
source = ["software/sdk/ternaryair"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
